portfolio-chatbot/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
httpx
beautifulsoup4